+++ app/src/main/java/com/mrgames13/jimdo/feinstaubapp/HelpClasses/DiagramEntry.kt
@@ -6,4 +6,4 @@ package com.mrgames13.jimdo.feinstaubapp.HelpClasses
 
 import com.github.mikephil.charting.data.Entry
 
-class DiagramEntry(x: Float, y: Float, internal val unit: String) : Entry(x, y)
+class DiagramEntry(x: Double, y: Double, internal val unit: String) : Entry(x.toFloat(), y.toFloat())
